feat: agentic search refactoring
