+++ src/main/kotlin/ai/masaic/openresponses/tool/agentic/SeedStrategy.kt
@@ -1,0 +1,13 @@
+package ai.masaic.openresponses.tool.agentic
+
+import ai.masaic.openresponses.api.model.Filter
+import ai.masaic.openresponses.api.model.VectorStoreSearchResult
+
+interface SeedStrategy {
+    suspend fun seed(
+        query: String,
+        maxResults: Int,
+        userFilter: Filter?,
+        vectorStoreIds: List<String>,
+    ): List<VectorStoreSearchResult>
+}
